Added Logic to add SSR in each flight

+++ src/main/java/com/AI_investigator/Controllers/FlightApi.java
@@ -50,7 +50,7 @@ public class FlightApi {
     }
 
     @PostMapping("/getFlightSSRs")
-    public ResponseEntity<List<SSR>> getFlightSSRs(@RequestBody int flightID, @RequestBody FareType fareType) {
+    public ResponseEntity<List<SSR>> getFlightSSRs(@RequestBody long flightID, @RequestBody FareType fareType) {
 
         List<SSR> ssrs;
